dev(hansbug): move CustomMapping for test.testings
