+++ test/testings/__init__.py
@@ -1,0 +1,1 @@
+from .mapping import CustomMapping
